Refactor the builder class
