+++ ParkingManagement/src/main/java/edu/du/ict4315/parkingmanagement/Main.java
@@ -1,7 +1,5 @@
 package edu.du.ict4315.parkingmanagement;
 
-import edu.du.ict4315.parkingmanagement.builder.customer.CustomerInfo;
-
 public class Main {
       public static void main(String[] args) {
             System.out.println("Hello world!");
