+++ ParkingManagement/src/test/java/edu/du/ict4315/parkingmanagement/RegisterCustomerCommandTest.java
@@ -4,7 +4,7 @@ package edu.du.ict4315.parkingmanagement; /**
  * @Instructor: Mike Prasad
  */
 
-import edu.du.ict4315.parkingmanagement.builder.address.AddressInfo;
+import edu.du.ict4315.parkingmanagement.builder.AddressInfo;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
